# openapi.yaml
openapi: 3.1.0
info:
  title: crypto-gateway
  version: 0.0.0
  description: |
    Runtime-agnostic crypto payment gateway with pluggable chain adapters
    (EVM / Tron / Solana), DB adapters (D1 / libSQL / Postgres), cache
    adapters, detection strategies, and webhook delivery.

    Deploys on Cloudflare Workers, Node/Bun, Deno, and Vercel Edge from one
    codebase.
servers:
  - url: http://localhost:8787
    description: Local Node dev
  - url: https://crypto-gateway-dev.workers.dev
    description: Dev Workers (placeholder — replace with your subdomain)
...
security:
  - MerchantApiKey: []
paths:
...
  /api/v1/invoices:
    post:
      tags: [invoices]
      summary: Create an invoice
      description: |
        Creates an invoice and allocates a receive address derived from the
        gateway's HD seed. The merchant is resolved from the API key — the
        caller MUST NOT include `merchantId` in the body; if they do, it's
        ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceInput" }
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                type: object
                required: [invoice]
                properties:
                  invoice: { $ref: "#/components/schemas/Invoice" }
          headers:
            X-Request-Id: { $ref: "#/components/headers/X-Request-Id" }
            X-RateLimit-Limit: { $ref: "#/components/headers/X-RateLimit-Limit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/X-RateLimit-Remaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/X-RateLimit-Reset" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/MerchantInactive" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/invoices/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [invoices]
      summary: Fetch an invoice
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                type: object
                required: [invoice]
                properties:
                  invoice: { $ref: "#/components/schemas/Invoice" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/invoices/{id}/expire:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [invoices]
      summary: Force-expire an invoice before its expiry window elapses
      responses:
        "200":
          description: Invoice expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice: { $ref: "#/components/schemas/Invoice" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Invoice is in a terminal state and cannot be expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    MerchantApiKey:
      type: http
      scheme: bearer
      description: |
        Merchant API key returned once at `POST /admin/merchants`.
        Also accepted as `X-API-Key: <key>`.
    AdminKey:
      type: http
      scheme: bearer
      description: Shared admin key from `ADMIN_KEY` env. Also accepted as `X-Admin-Key`.
    AlchemySignature:
      type: apiKey
      in: header
      name: X-Alchemy-Signature
      description: HMAC-SHA256 hex of the raw body under the per-webhook signing key stored in `alchemy_webhook_registry` (looked up by `event.webhookId`).
    CronSecret:
      type: http
      scheme: bearer
      description: Shared secret from `CRON_SECRET` env.
  headers:
    X-Request-Id:
      description: UUID correlation id. Accepts a caller-supplied value if it's a valid UUID; otherwise minted.
      schema: { type: string, format: uuid }
    X-RateLimit-Limit:
      description: Window cap for this route's rate-limit scope
      schema: { type: integer }
    X-RateLimit-Remaining:
      description: Requests left in the current window after this one
      schema: { type: integer }
    X-RateLimit-Reset:
      description: Unix seconds when the window rolls
      schema: { type: integer }
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    MerchantInactive:
      description: Merchant is deactivated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Entity not found (or not owned by this merchant)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests in the current window
      headers:
        Retry-After:
          description: Seconds until the window rolls
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code]
          properties:
            code: { type: string, description: Stable machine-readable code }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Optional per-code payload (e.g. Zod issues on VALIDATION)
    CreateInvoiceInput:
      type: object
      description: |
        Either `amountRaw` OR (`fiatAmount` + `fiatCurrency`) must be provided.
      properties:
        chainId: { type: integer }
        token: { type: string, pattern: "^[A-Z0-9]+$" }
        amountRaw:
          type: string
          pattern: "^\\d+$"
          description: Raw token amount in smallest units.
        fiatAmount:
          type: string
          pattern: "^\\d+(\\.\\d+)?$"
        fiatCurrency:
          type: string
          pattern: "^[A-Z]{3}$"
        externalId:
          type: string
          maxLength: 256
          description: Merchant's opaque reference (e.g. cart id)
        metadata:
          type: object
          additionalProperties: true
        expiresInMinutes:
          type: integer
          minimum: 1
          maximum: 1440
          default: 30
      required: [chainId, token]
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        status:
          type: string
          enum: [created, partial, detected, confirmed, expired, canceled]
        chainId: { type: integer }
        token: { type: string }
        receiveAddress: { type: string }
        addressIndex: { type: integer }
        requiredAmountRaw: { type: string }
        receivedAmountRaw: { type: string }
        fiatAmount: { type: [string, "null"] }
        fiatCurrency: { type: [string, "null"] }
        quotedRate: { type: [string, "null"] }
        externalId: { type: [string, "null"] }
        metadata: { type: [object, "null"] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        confirmedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
...